Time taken to create 3276700 instances of FiveCharacterString: 2212 ms
