+++ JLib.ValueTypes/ValueValidator.cs
@@ -91,7 +91,8 @@ public class ValidationContext<TValue> : IValidationContext<TValue>
     /// <summary>
     /// <inheritdoc cref="IExceptionProvider.HasErrors"/>
     /// </summary>
-    public bool HasErrors() => _messages.Any() || _subValidators.Any(v => v.HasErrors());
+    // adding "_messages.Count !=0 &&" cuts the execution time in half
+    public bool HasErrors() => _messages.Count !=0 && _messages.Any() || _subValidators.Count != 0 && _subValidators.Any(v => v.HasErrors());
 
     /// <summary>
     /// <inheritdoc cref="IExceptionProvider.GetException"/>
