+++ Tools/JLib.Tools.Benchmarks/PerfTest.cs
@@ -12,16 +12,16 @@ using System.Reflection;
 
 namespace JLib.Tools.Benchmarks;
 
-public class PerfTest
+[InProcess]
+public partial class PerfTest
 {
+    private static readonly Random __random = new Random();
     private static readonly Func<string> _getRandom5LetterString = GetNextString;
 
     //[Config(typeof(Config))]
     private static readonly List<string> _strings = Enumerable.Range(0, 5000).Select(_ => GetRandom5LetterString()).ToList();
 
     private static IEnumerator<string> _stringEnumerator = _strings.GetEnumerator();
-    private static Random random = new Random();
-
     public record FiveCharacterString(string Value) : StringValueType(Value)
     {
         [Validation]
@@ -40,7 +40,7 @@ public class PerfTest
 
     public static string GetRandom5LetterString()
     {
-        var str = random.Next(10000, 99999).ToString();
+        var str = __random.Next(10000, 99999).ToString();
         return $"{str}";
     }
 
@@ -55,6 +55,8 @@ public class PerfTest
         }
         return _stringEnumerator.Current;
     }
+
+
 
     public class CheckPerf
     {
@@ -81,9 +83,16 @@ public class PerfTest
             WithOptions(ConfigOptions.DisableOptimizationsValidator);
         }
     }
-
+    [Config(typeof(Config))]
+    [InProcess]
     public class StringPerformanceBenchmarks
     {
+
+        [Benchmark]
+        public string FiveCharacterString()
+        {
+            return FiveCharacterStringCreate(_getRandom5LetterString).Value;
+        }
         [Benchmark]
         public string ConcatenationOperator()
         {
@@ -100,60 +109,6 @@ public class PerfTest
         public string StringInterpolation()
         {
             return $"{_getRandom5LetterString()}{_getRandom5LetterString()}{_getRandom5LetterString()}{_getRandom5LetterString()}{_getRandom5LetterString()}";
-        }
-    }
-
-    // Crimes against god here
-    public class Swapper
-    {
-        // Create a Harmony instance
-        private static Harmony harmony = new Harmony("com.example.patch");
-
-        // Get the method to patch
-        private static MethodInfo originalMethod = typeof(ExceptionBuilder)!.GetMethod("GetException")!;
-
-        [Params("patch", "lockedpatch", "unpatched")]
-        public string? PatchState { get; set; }
-
-        [Benchmark]
-        public string FiveCharacterString()
-        {
-            return FiveCharacterStringCreate(_getRandom5LetterString).Value;
-        }
-
-        [GlobalSetup(Target = nameof(FiveCharacterString))]
-        public void GlobalSetup2()
-        {
-            harmony.Unpatch(originalMethod, HarmonyPatchType.All, "com.example.patch");
-            if (PatchState == "patch")
-            {
-                harmony.Patch(originalMethod, new HarmonyMethod(typeof(PatchClass).GetMethod(nameof(PatchClass.Prefix))!));
-            }
-            if (PatchState == "lockedpatch")
-            {
-                harmony.Patch(originalMethod, new HarmonyMethod(typeof(PatchClass).GetMethod(nameof(PatchClass.LockedPrefix))!));
-            }
-        }
-
-        // #region Experimental Patching
-        public class PatchClass
-        {
-            public static bool LockedPrefix(List<Exception> ____exceptions, List<IExceptionProvider> ____children, object ____exceptionLock, object ____childrenLock)
-            {
-                lock (____exceptionLock)
-                    lock (____childrenLock)
-                        return Prefix(____exceptions, ____children);
-            }
-
-            public static bool Prefix(List<Exception> ____exceptions, List<IExceptionProvider> ____children)
-            {
-                bool condition = ____exceptions.Count == 0 && ____children.Count == 0;
-                if (condition)
-                {
-                    return false; // Returning false skips the original method
-                }
-                return true; // Returning true allows the original method to execute
-            }
         }
     }
 }
